+++ shoulder-build/shoulder-base/shoulder-web/src/main/java/org/shoulder/web/template/dictionary/model/DictionaryItemEntity.java
@@ -59,7 +59,7 @@ public class DictionaryItemEntity<ID extends Serializable> extends BizEntity<ID>
     }
 
     @Override
-    public long getDisplayOrder() {
+    public Integer getDisplayOrder() {
         return sortNo;
     }
